fix: [TEO-250] Move lambda "preauthentication check" from serverless to terraform
