--- lambda/pre-auth-permissions-check/.prettierrc
@@ -1,5 +1,0 @@
-{
-    "singleQuote": true,
-    "trailingComma": "all",
-    "tabWidth": 2
-}
