--- lambda/pre-auth-permissions-check/serverless.yml
@@ -1,12 +1,0 @@
-service: auth-hook
-frameworkVersion: '3'
-
-provider:
-  name: aws
-  runtime: nodejs14.x
-  region: eu-central-1
-  stage: dev
-
-functions:
-  permissionCheck:
-    handler: handler.permissionCheck
